created PlayerCard component
